test: move matchers to expect.spec.ts and remove fixme
